Adds first commit - testing framework for executables - bash test tools
A framework for testing executables in a bash shell environment.
Intended to harmonize, simplify and avoid hard to read shell scripts for
testing. Developed as part of a CI and QA at Videntifier Technologies.

Documentation can be compiled with mkdocs.

TODO: Add actual test for the test framework.
TODO: Add more test examples